reduce code

+++ src/main/java/com/example/final_case_social_web/controller/AdminRestController.java
@@ -33,8 +33,9 @@ public class AdminRestController {
     private PostService postService;
 
     // Xem tất cả user
-    @GetMapping("/getAllUser")
-    public ResponseEntity<?> getAllUser(@RequestParam Long idAdmin,
+    @GetMapping("/adminAction")
+    public ResponseEntity<?> adminAction(@RequestParam Long idAdmin,
+                                        @RequestParam String type,
                                         @RequestHeader("Authorization") String authorization) {
         if (userService.checkAdmin(idAdmin) == null) {
             return new ResponseEntity<>(ResponseNotification.
@@ -44,16 +45,25 @@ public class AdminRestController {
         if (responseEntity != null) {
             return responseEntity;
         }
-        List<User> users = userService.findAllRoleUser();
-        List<UserDTO> userDTOList = new ArrayList<>();
-        if (!CollectionUtils.isEmpty(users)) {
-            for (User user : users) {
-                UserDTO userDTO = new UserDTO();
-                BeanUtils.copyProperties(user, userDTO);
-                userDTOList.add(userDTO);
+        if ("user".equals(type)) {
+            List<User> users = userService.findAllRoleUser();
+            List<UserDTO> userDTOList = new ArrayList<>();
+            if (!CollectionUtils.isEmpty(users)) {
+                for (User user : users) {
+                    UserDTO userDTO = new UserDTO();
+                    BeanUtils.copyProperties(user, userDTO);
+                    userDTOList.add(userDTO);
+                }
             }
+            return new ResponseEntity<>(userDTOList, HttpStatus.OK);
         }
-        return new ResponseEntity<>(userDTOList, HttpStatus.OK);
+        if ("post".equals(type)) {
+            Iterable<Post2> post2List = postService.findAll();
+            List<Post2> list = (List<Post2>) post2List;
+            list.sort((p1, p2) -> (p2.getCreateAt().compareTo(p1.getCreateAt())));
+            return new ResponseEntity<>(list, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @GetMapping("/findById")
@@ -66,10 +76,11 @@ public class AdminRestController {
         return new ResponseEntity<>(userOptional.get(), HttpStatus.OK);
     }
 
-    // Cấm user
-    @DeleteMapping("/changeStatusUserBan")
-    public ResponseEntity<?> changeStatusUserBan(@RequestParam Long idAdmin,
+    // Cấm user, Kích hoạt tài khoản
+    @DeleteMapping("/actionUser")
+    public ResponseEntity<?> actionUser(@RequestParam Long idAdmin,
                                                  @RequestParam Long idUser,
+                                                 @RequestParam String type,
                                                  @RequestHeader("Authorization") String authorization) {
         if (userService.checkAdmin(idAdmin) == null) {
             return new ResponseEntity<>(ResponseNotification.
@@ -85,7 +96,12 @@ public class AdminRestController {
                     responseMessage(Constants.IdCheck.ID_USER, idUser), HttpStatus.NOT_FOUND);
         }
         if (userService.checkAdmin(idAdmin).toString().substring(17, 27).equals(Constants.Roles.ROLE_ADMIN)) {
-            optionalUser.get().setStatus(Constants.STATUS_BANED);
+            if ("baned".equals(type)) {
+                optionalUser.get().setStatus(Constants.STATUS_BANED);
+            }
+            if ("active".equals(type)) {
+                optionalUser.get().setStatus(Constants.STATUS_ACTIVE);
+            }
             userService.save(optionalUser.get());
             return new ResponseEntity<>(HttpStatus.OK);
         }
@@ -94,38 +110,11 @@ public class AdminRestController {
                 HttpStatus.BAD_REQUEST);
     }
 
-    // Kích hoạt tài khoản
-    @DeleteMapping("/changeStatusActive")
-    public ResponseEntity<?> changeStatusActive(@RequestParam Long idAdmin,
-                                                 @RequestParam Long idUser,
-                                                 @RequestHeader("Authorization") String authorization) {
-        if (userService.checkAdmin(idAdmin) == null) {
-            return new ResponseEntity<>(ResponseNotification.
-                    responseMessage(Constants.IdCheck.ID_ADMIN, idAdmin), HttpStatus.UNAUTHORIZED);
-        }
-        ResponseEntity<?> responseEntity = userService.errorToken(authorization, idAdmin);
-        if (responseEntity != null) {
-            return responseEntity;
-        }
-        Optional<User> optionalUser = userService.findById(idUser);
-        if (!optionalUser.isPresent()) {
-            return new ResponseEntity<>(ResponseNotification.
-                    responseMessage(Constants.IdCheck.ID_USER, idUser), HttpStatus.NOT_FOUND);
-        }
-        if (userService.checkAdmin(idAdmin).toString().substring(17, 27).equals(Constants.Roles.ROLE_ADMIN)) {
-            optionalUser.get().setStatus(Constants.STATUS_ACTIVE);
-            userService.save(optionalUser.get());
-            return new ResponseEntity<>(HttpStatus.OK);
-        }
-        return new ResponseEntity<>(new ResponseNotification(HttpStatus.BAD_REQUEST.toString(),
-                MessageResponse.NO_VALID, MessageResponse.DESCRIPTION),
-                HttpStatus.BAD_REQUEST);
-    }
-
-    // Xoá bài viết trong database
-    @DeleteMapping("/deletePost")
-    public ResponseEntity<?> deletePost(@RequestParam Long idAdmin,
+    // Xoá bài viết trong database, khóa bài viết
+    @DeleteMapping("/actionPost")
+    public ResponseEntity<?> actionPost(@RequestParam Long idAdmin,
                                         @RequestParam Long idPost,
+                                        @RequestParam String type,
                                         @RequestHeader("Authorization") String authorization) {
         if (userService.checkAdmin(idAdmin) == null) {
             return new ResponseEntity<>(ResponseNotification.
@@ -141,56 +130,17 @@ public class AdminRestController {
                     responseMessage(Constants.IdCheck.ID_POST, idPost), HttpStatus.NOT_FOUND);
         }
         if (userService.checkAdmin(idAdmin).toString().substring(17, 27).equals(Constants.Roles.ROLE_ADMIN)) {
-            postService.delete(postOptional.get());
+            if ("delete".equals(type)) {
+                postService.delete(postOptional.get());
+            }
+            if ("lock".equals(type)) {
+                postOptional.get().setStatus(Constants.STATUS_PRIVATE);
+                postService.save(postOptional.get());
+            }
             return new ResponseEntity<>(HttpStatus.OK);
         }
         return new ResponseEntity<>(new ResponseNotification(HttpStatus.BAD_REQUEST.toString(),
                 MessageResponse.NO_VALID, MessageResponse.DESCRIPTION),
                 HttpStatus.BAD_REQUEST);
-    }
-
-    // Khóa bài viết
-    @DeleteMapping("/lockPost")
-    public ResponseEntity<?> lockPost(@RequestParam Long idAdmin,
-                                      @RequestParam Long idPost,
-                                      @RequestHeader("Authorization") String authorization) {
-        if (userService.checkAdmin(idAdmin) == null) {
-            return new ResponseEntity<>(ResponseNotification.
-                    responseMessage(Constants.IdCheck.ID_ADMIN, idAdmin), HttpStatus.UNAUTHORIZED);
-        }
-        ResponseEntity<?> responseEntity = userService.errorToken(authorization, idAdmin);
-        if (responseEntity != null) {
-            return responseEntity;
-        }
-        Optional<Post2> postOptional = postService.findById(idPost);
-        if (!postOptional.isPresent()) {
-            return new ResponseEntity<>(ResponseNotification.
-                    responseMessage(Constants.IdCheck.ID_POST, idPost), HttpStatus.NOT_FOUND);
-        }
-        if (userService.checkAdmin(idAdmin).toString().substring(17, 27).equals(Constants.Roles.ROLE_ADMIN)) {
-            postOptional.get().setStatus(Constants.STATUS_PRIVATE);
-            postService.save(postOptional.get());
-            return new ResponseEntity<>(HttpStatus.OK);
-        }
-        return new ResponseEntity<>(new ResponseNotification(HttpStatus.BAD_REQUEST.toString(),
-                MessageResponse.NO_VALID, MessageResponse.DESCRIPTION),
-                HttpStatus.BAD_REQUEST);
-    }
-
-    @GetMapping("/allPost")
-    public ResponseEntity<?> allPost(@RequestParam Long idAdmin,
-                                     @RequestHeader("Authorization") String authorization) {
-        if (userService.checkAdmin(idAdmin) == null) {
-            return new ResponseEntity<>(ResponseNotification.
-                    responseMessage(Constants.IdCheck.ID_ADMIN, idAdmin), HttpStatus.UNAUTHORIZED);
-        }
-        ResponseEntity<?> responseEntity = userService.errorToken(authorization, idAdmin);
-        if (responseEntity != null) {
-            return responseEntity;
-        }
-        Iterable<Post2> post2List = postService.findAll();
-        List<Post2> list = (List<Post2>) post2List;
-        list.sort((p1, p2) -> (p2.getCreateAt().compareTo(p1.getCreateAt())));
-        return new ResponseEntity<>(list, HttpStatus.OK);
     }
 }
